actualizacion

+++ vueltaAjava/Tema03/OtraVez02.java
@@ -5,11 +5,9 @@ public class OtraVez02{
   public static void main(String args[]){
   System.out.print("este programa realiza un cambio de euros a pesetas" + "\nintroduce una cantidad en euros\n");
   
-  int euros = Integer.parseInt(System.console().readLine());
+  double euros = Double.parseDouble(System.console().readLine());
   double conversor = (euros*166.6);
    
    System.out.println(euros + " euros son " + "\n"+conversor + " pesetas");
-  
-  
   }
 }
